added styled prompt support
