+++ docker/server/index.js
@@ -65,6 +65,7 @@ app.post("/config", (req, res) => {
     summary_num_results,
     summary_num_sentences,
     summary_prompt_name,
+    summary_styled_prompt,
 
     // rerank
     rerank,
@@ -130,7 +131,8 @@ app.post("/config", (req, res) => {
     summary_num_results,
     summary_num_sentences,
     summary_prompt_name,
-
+    summary_styled_prompt,
+    
     // hybrid search
     hybrid_search_num_words,
     hybrid_search_lambda_long,
