+++ server/index.js
@@ -64,6 +64,7 @@ app.post("/config", (req, res) => {
     summary_num_results,
     summary_num_sentences,
     summary_prompt_name,
+    summary_styled_prompt,
 
     // hybrid search
     hybrid_search_num_words,
@@ -129,6 +130,7 @@ app.post("/config", (req, res) => {
     summary_num_results,
     summary_num_sentences,
     summary_prompt_name,
+    summary_styled_prompt,
 
     // hybrid search
     hybrid_search_num_words,
